More robust error handling
Added more robust error handling - after a loop (to be replaced by a check on loop errors once Discord.py v1.4 is released), we check why the loop was cancelled (since it should never be cancelled). If it failed or if it was explicilty cancelled, we send a message to the health channels and attempt to cancel + restart the loop task.