Added in Post Removal capabilities
If a filter's action is to remove content, it now actually does so. Subreddit lists also have a flag set by the user to specify whether they have moderator access to remove posts. This is currently separate from the flag to edit automoderator configs and will need to be consolidated.